+++ lib/src/property_serializer.ts
@@ -1,4 +1,4 @@
-import { SerializationOptions } from "./options";
+import { DeflateOrInflateOptions } from "./options";
 
 /** Reporesents a generic property serializer */
 interface PropertySerializer<TSerialized, TOriginal, TTag> {
@@ -10,10 +10,10 @@ interface PropertySerializer<TSerialized, TOriginal, TTag> {
     readonly propertyTag: TTag;
 
     /** Serializes target property from `serializable` and writes value to `serialized` */
-    down(serializable: TOriginal, serialized: TSerialized, options?: SerializationOptions): void;
+    down(serializable: TOriginal, serialized: TSerialized, options?: DeflateOrInflateOptions<TSerialized, TOriginal>): void;
 
     /** Deserializes target property from `serialized` and writes value to `serializable` */
-    up(serializable: TOriginal, serialized: TSerialized, options?: SerializationOptions): void;
+    up(serializable: TOriginal, serialized: TSerialized, options?: DeflateOrInflateOptions<TSerialized, TOriginal>): void;
 
 }