+++ lib/src/type_serializer_picker.ts
@@ -1,12 +1,12 @@
 import { DEFAULT_PROJECTION, MetadataManager } from './metadata';
-import { SerializationOptions } from './options';
+import { DeflateOrInflateOptions } from './options';
 import { Constructor, isConstructor } from './types/constructor';
 import TypeSerializer from './type_serializer';
 
 /** Returns a helper which picks a type serializer for given value or type */
 export default function TypeSerializerPicker<TSerialized, TOriginal>(
     backend: string,
-    options?: SerializationOptions
+    options?: DeflateOrInflateOptions<TSerialized, TOriginal>
 ) {
 
     let { projection, fallbackToDefaultProjection = true } = options || {};